remove fixmeharrison
