+++ streaming/csharp/compass/.DS_Store
