+++ streaming/csharp/compass/harvest/v1alpha1/Ingest.cs
