+++ streaming/csharp/compass/harvest/v1alpha1/IngestGrpc.cs
