+++ streaming/csharp/compass/iam/Permission.cs
