+++ streaming/csharp/compass/type/geo/Geo.cs
